+++ src/Graph/Nodes/Isolator.php
@@ -69,7 +69,9 @@ class Isolator implements HasBinaryChildren
         if ($type->isContinuous()) {
             [$min, $max] = minmax($values);
 
-            $phi = getrandmax() / max(abs($max), abs($min));
+            $maxAbs = max(abs($max), abs($min));
+
+            $phi = $maxAbs > 0.0 ? getrandmax() / $maxAbs : getrandmax();
 
             $min = (int) floor($min * $phi);
             $max = (int) ceil($max * $phi);