+++ src/Transformers/RobustStandardizer.php
@@ -184,6 +184,6 @@ class RobustStandardizer implements Transformer, Stateful, Reversible, Persistab
      */
     public function __toString() : string
     {
-        return 'Robust Standardizer {center: ' . Params::toString($this->center) . ')';
+        return 'Robust Standardizer (center: ' . Params::toString($this->center) . ')';
     }
 }