+++ tests/Strategies/WildGuessTest.php
@@ -67,4 +67,16 @@ class WildGuessTest extends TestCase
             )
         );
     }
+
+    /**
+     * @test
+     */
+    public function fitGuessConstantZero() : void
+    {
+        $this->strategy->fit([0.0, 0.0, 0.0]);
+
+        $this->assertTrue($this->strategy->fitted());
+
+        $this->assertEquals(0.0, $this->strategy->guess());
+    }
 }