+++ tests/Extractors/ColumnPickerTest.php
@@ -43,4 +43,26 @@ class ColumnPickerTest extends TestCase
 
         $this->assertEquals($expected, $records);
     }
+
+    /**
+     * @test
+     */
+    public function extractNullColumn() : void
+    {
+        $iterable = (function () {
+            yield [
+                'attitude' => 'nice', 'texture' => null, 'class' => 'not monster', 'rating' => '4',
+            ];
+        })();
+
+        $extractor = new ColumnPicker($iterable, ['texture']);
+
+        $expected = [
+            ['texture' => null],
+        ];
+
+        $records = iterator_to_array($extractor, false);
+
+        $this->assertEquals($expected, $records);
+    }
 }