+++ src/Report.php
@@ -12,6 +12,8 @@ use Traversable;
 use Stringable;
 use Countable;
 
+use function array_key_exists;
+
 /**
  * Report
  *
@@ -80,7 +82,7 @@ class Report implements ArrayAccess, JsonSerializable, IteratorAggregate, Counta
      */
     public function offsetExists($key) : bool
     {
-        return isset($this->attributes[$key]);
+        return array_key_exists($key, $this->attributes);
     }
 
     /**
@@ -93,7 +95,7 @@ class Report implements ArrayAccess, JsonSerializable, IteratorAggregate, Counta
     #[\ReturnTypeWillChange]
     public function offsetGet($key)
     {
-        if (isset($this->attributes[$key])) {
+        if (array_key_exists($key, $this->attributes)) {
             return $this->attributes[$key];
         }