+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     - Fix SQL Table offset syntax to work with Postgres
     - Prevent SQL Table injection attack
     - Pipeline add missing traided guard on score method
+    - Fix Column Picker/Filter on null columns
+    - Fix Image Rotator mangling non-90° rotations
 
 - 2.5.12
     - Fix Adaboost proba() probability normalization