+++ CHANGELOG.md
@@ -46,7 +46,11 @@
     - `Xavier 2` now extends `He` as a backward-compatible alias
     
 - 2.5.13
-    - Pipeline add missing traided guard on score method
+    - Fix SQL Table offset syntax to work with Postgres
+    - Prevent SQL Table injection attack
+    - Pipeline add missing trained guard on score method
+    - Fix Column Picker/Filter on null columns
+    - Fix Image Rotator mangling non-90° rotations
 
 - 2.5.12
     - Fix Adaboost proba() probability normalization